+++ src/main/java/com/woorifisa/backend/ex/MemberService.java
@@ -1,36 +1,11 @@
 package com.woorifisa.backend.ex;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-import jakarta.transaction.Transactional;
+public interface MemberService {
 
-@Service
-@RequiredArgsConstructor
-public class MemberService {
+    public MemberDTO getMemberById(Long id);
 
+    public void updateMemberName(Long id, String name);
 
-    private final MemberRepository memberRepository;
-
-
-    public MemberDTO getMemberById(Long id) {
-        MemberEx member = memberRepository.findById(id).orElse(null);
-        System.out.println(member);
-        return MemberDTO.builder()
-                        .id(member.getId())
-                        .Birth(member.getBirth())
-                        .name(member.getName())
-                        .email(member.getEmail())
-                        .build();
-    }
-    
-
-    // save() 를 통한 update
-    @Transactional
-    public void updateMemberName(Long id, String name){
-        MemberEx member = memberRepository.findById(id).orElse(null);
-        member.setName(name);
-        memberRepository.save(member);
-    }
 }
